Primary Calculator Logic & CalculationResult Class
This commit was delayed mainly because I had gone back to adjust the methods to accommodate for the CalculationResult class, which makes the code much more testable, as advised by my professor Jeff Adkisson. All of the functionalities of the Calculator Logic is entered here with the correct outputs. Though, the Population Standard Deviation and the single linear regression does not 100% match the requirements as written. Documentation will be added and the computeSingleRegression method could be modularized to more easily test the function later on.
The main focus of the next week will be creating all of the unit tests necessary, which should not take long. Only after having 100% coverage will the webserver and end to end testing be done with the end to end tests being the finale of the project.