Demo WebServer - Non-functional but the framework has been set
Here there are few changes made. The main focus on this and the forthcoming update is the creation of a functional webserver with an interface to match the project requirements. I am working hard to understand spring and thymeleaf, but it will take some time. The web page can post with a simple text to multiple pages, but additional functionalities will be added in the coming days.

Goals are to make the program executable through command line interface, implementing functionalities through spring and thymeleaf, and integrating these with the Calculator Functions module to output the correct data.

Later on, the end to end testing will be done, but that is at the tail end of the development process.