Semi-final Commit - All functions operational with proper input
Proper form handling and functions are operational with Spring boot and Thymeleaf for variable handling across from Java environment to the HTML templates used for modeling on the webpage. There are still tests to be performed. However, with the appropriate input to the textbox, all the functions are capable of generating the appropriate return, and errors will be shown on page. Use Clear button to return to the initial page. Thus, the concept design has been confirmed, and the project is close to completion though it lacks refinement.
This has been a major learning experience for front end designing as well. This portion of the project was much more involved than expected, but thymeleaf and spring has proven itself to be extremely useful for these purposes