Semi-Final Commit - Will Receive Refinement and More
This comes as a result of updating my operating system; thus, the build failed right before I was to submit this project, so I had to carry the work to the next day. This result has caused anguish and confusion.

The final addition to the project will be the end to end testing which will be completed after proper configuration has been applied for complete operation of the other 3 modules. Major error on the Mac OS System was a failure to find the html templates which was used to model the webpages returned through thymeleaf. This operation is function on Windows, so progression will likely continue there.
This Commit is just for the purpose of submission, as the project has enough to get a passing grade. Further improvements will still be made even if perfection is not possible